+++ packages/core-js/full/iterator/index.js
@@ -2,12 +2,10 @@
 var parent = require('../../actual/iterator');
 require('../../modules/esnext.iterator.chunks');
 require('../../modules/esnext.iterator.range');
-require('../../modules/esnext.iterator.sliding');
 require('../../modules/esnext.iterator.windows');
-require('../../modules/esnext.iterator.zip');
-require('../../modules/esnext.iterator.zip-keyed');
 // TODO: Remove from `core-js@4`
 require('../../modules/esnext.iterator.as-indexed-pairs');
 require('../../modules/esnext.iterator.indexed');
+require('../../modules/esnext.iterator.sliding');
 
 module.exports = parent;